requirements.txt
aiogram==3.8.0
gspread==5.12.4
oauth2client==4.1.3
python-dotenv==1.0.0
pytz==2024.1
google-api-python-client==2.100.0